Initial updates to the nav bar
