+++ src/Components/Navigation/Components/index.js
@@ -1,2 +1,3 @@
 export { default as NavMini } from './NavMini';
 export { default as NavStandard } from './NavStandard';
+export { default as NavHeaders } from './NavHeaders';
