+++ src/Pages/Gas/Gas.js
@@ -3,7 +3,7 @@ import { Section, Wrapper } from 'Components';
 import { GasVolume, GasStats } from './Components';
 
 const Gas = () => (
-  <Wrapper>
+  <Wrapper noHeader>
     <Section>
       <GasStats />
     </Section>
